Move tpls guide to component and wrap field names
